refactor(hal): change names of library to time_interface and Time
BREAKING CHANGE: change interface name
