+++ src/hal/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_subdirectory(gpio)
 add_subdirectory(sleep)
-add_subdirectory(timeFacade)
+add_subdirectory(time)
 
 add_subdirectory(i2c)
 add_subdirectory(spi)
