+++ src/hal/time/Time.c
@@ -1,6 +1,7 @@
-#include "TimeFacade.h"
+#include "Time.h"
 
 #include "pico/time.h"
+#include <stdint.h>
 
 uint64_t get_current_time_in_us(void) {
     return (uint64_t)get_absolute_time();
