+++ test/unit/dummies/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_subdirectory(sleep)
-add_subdirectory(timeFacade)
+add_subdirectory(time)
 add_subdirectory(i2c)
 add_subdirectory(gpio)
 add_subdirectory(esp)
