+++ test/unit/dummies/time/Time.c
@@ -1,6 +1,6 @@
 #define SOURCE_FILE "DUMMY-TIME-FACADE-LIB"
 
-#include "TimeFacade.h"
+#include "Time.h"
 #include "Common.h"
 
 uint64_t get_current_time_in_us(void) {
